Ensure thumbnails match document img order and link them.
Make a click on the thumbnail take you to the larger image
in the context of the post. (We do this assigning an id attribute
to each img, and taking advantage of the fact the browser will
scroll the element with ID into view for URL "#" + ID.)

Also drop the limit on the number of thumbnails.